updated the docker-compose file for the databases, added a folder for javascript functions(checkValidToken, createAdminUser) that are used inside the routs and app.js
