+++ DataBases/docker-compose.yml
@@ -1,7 +1,7 @@
 version: "3"
 
 services:
-  mongodb:
+  mongodb_auth:
     image: mongo:5.0.16-focal
     container_name: mongodb-auth
     # environment:
@@ -10,9 +10,17 @@ services:
     ports:
       - 4001:27017
     # volumes:
-    #   - /home/falape/Nearsoft/Docker/dremio-env:/opt/dremio/conf/dremio-env
+    #   - /tmp/mongo-data:/data/db mongo:6-jammy
     restart: unless-stopped
 
-  
-# -e MONGO_INITDB_DATABASE=test 
-# -v /tmp/mongo-data:/data/db mongo:6-jammy
+  mongodb_data:
+    image: mongo:5.0.16-focal
+    container_name: mongodb-data-api
+    # environment:
+    #   - MONGO_INITDB_ROOT_USERNAME=admin
+    #   - MONGO_INITDB_ROOT_PASSWORD=12345
+    ports:
+      - 4003:27017
+    # volumes:
+    #   - /tmp/mongo-data:/data/db mongo:6-jammy
+    restart: unless-stopped
